addressing PR comments
